fix(renderer): correct PDH API misuse, GPU% precision, and memory metric
- PdhGpuMonitor: PdhGetFormattedCounterValue 3rd param is counter TYPE code
  (PERF_100NSEC_TIMER), not status; checking == ERROR_SUCCESS made all 15
  engine counters report invalid. Pass nullptr, check return value only.
- IGpuMonitor: gpuUtilizationPercent uint32_t -> double (0.9% not 1.0%)
- ProcessStatsCollector: WorkingSetSize -> PrivateWorkingSetSize via
  PROCESS_MEMORY_COUNTERS_EX2 (matches Task Manager, excludes shared DLLs)

+++ renderer/src/monitor/IGpuMonitor.hpp
@@ -15,7 +15,7 @@ namespace Monitor {
  * byte-for-byte consistent with controller model/RendererStats.java.
  */
 struct GpuMetrics {
-    std::optional<uint32_t> gpuUtilizationPercent;
+    std::optional<double> gpuUtilizationPercent;
     std::optional<uint64_t> vramUsedBytes;
     std::optional<uint64_t> vramTotalBytes;
     std::string gpuName;
